+++ libfxcg/misc/calloc.c
@@ -2,7 +2,7 @@
 
 extern void* memsetZero(void *dest, int n);
 
-void *calloc(int elements, int elementSize){
+void *calloc(size_t elements, size_t elementSize){
     int p = elements*elementSize;
     void *tmp;
     tmp=malloc(p);
